New sector implementation.
This removes the list of warps for a sector, and instead each sector has a warp in each cardinal direction on a cartesian plane (z axis is present, but it is always 0).

It should be easier to generate sectors now, as coordinates will define what sector a player is in, and a name can be generated for each sector to make it unique, but is not necessary for this to work.

The sector lookup (Game._sectors) now uses Coordinates as the key, but this is not saved in the Sector object, so it is added in the json that is returned to the client.