Fix Game.state() so it returns a dictionary.
Before, when it was converting to json, it would return the User object as a string, so {'user': "{'name':...}"}, which was not being handled on the client side (and converted to a dictionary from json). This adds the stipulation that Game.state() must provide a complete dictionary, and any objects must be converted to a dictionary before they can be returned.

The dictionary is returned in Base.to_dict(). This can be overridden, as in Ship